LayoutTests/crypto/workers/subtle/ecdsa-import-key-sign-expected.txt

[Worker] Test ECDSA signing operation with a P-256 private key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ecdsa-import-key-sign.js
PASS [Worker] signature.byteLength is 64
PASS successfullyParsed is true

TEST COMPLETE
